Split large file into three parts
